+++ optimus/pages/views/__init__.py
@@ -1,0 +1,8 @@
+from .base import PageViewBase
+from .template import PageTemplateView
+
+
+__all__ = [
+    "PageViewBase",
+    "PageTemplateView",
+]