+++ optimus/pages/views/base.py
@@ -1,21 +1,17 @@
-# -*- coding: utf-8 -*-
 import logging
 import os
 
 from jinja2 import meta as Jinja2Meta
 
-from optimus.exceptions import ViewImproperlyConfigured
-from optimus.i18n.lang import LangBase
+from ..exceptions import ViewImproperlyConfigured
+from ..i18n.lang import LangBase
 
 
 class PageViewBase:
     """
-    Base view object for a page
+    Base view object for a page.
 
-    You can set class attributes at the init if needed
-
-    The render method is responsible to rendering the HTML from the template
-    and his context. Actually this is the only used method directly.
+    You can set class attributes at the init if needed.
 
     Only ``lang`` and ``context`` attributes are optional, so take care to set
     all the required ones because their default value is ``None``. You should
@@ -30,8 +26,6 @@ class PageViewBase:
         Page destination
     page_lang
         Given langage if any
-    page_template_name
-        Template name used to compile the page HTML
 
     But you can add new variable if needed. The default context variables can
     not be overriden from the ``context`` class attribute, only from the
@@ -42,15 +36,11 @@ class PageViewBase:
 
     Attributes:
         title (string): Page title.
-        template_name (string): Page template file path relaive to templates
-            directoy. Used as Python template string with optional non
-            positional argument ``{{ language_code }}`` available for
-            internationalized pages.
         destination (string): Page destionation path relative to build
             directory.
         lang (string): Language identifier or an instance of
             ``optimus.i18n.LangBase``.
-        context (dict): Initial page template context.
+        context (dict): Initial page view context.
         logger (logging.Logger): Optimus logger.
         _used_templates (list): List of every used templates. Only filled when
             ``introspect()`` method is executed. Default to ``None``.
@@ -63,10 +53,10 @@ class PageViewBase:
     """
 
     title = None
-    template_name = None
     destination = None
     lang = None
     context = {}
+    _required_page_attributes = ["title", "destination"]
 
     def __init__(self, **kwargs):
         self._used_templates = None
@@ -100,11 +90,11 @@ class PageViewBase:
         Returns:
             boolean: ``True`` if requirements are set.
         """
-        err = []
-
-        for item in ["title", "template_name", "destination"]:
-            if not getattr(self, item):
-                err.append(item)
+        err = [
+            item
+            for item in self._required_page_attributes
+            if not getattr(self, item)
+        ]
 
         if len(err) > 0:
             msg = "These attributes are required: {}".format(", ".join(err))
@@ -124,7 +114,7 @@ class PageViewBase:
         """
         if not self.__settings:
             msg = (
-                """View required settings defined either from init """
+                """View requires settings defined either from init """
                 """arguments or through settings attribute"""
             )
             raise ViewImproperlyConfigured(msg)
@@ -144,7 +134,7 @@ class PageViewBase:
         """
         Get page title.
 
-        Default behavior is to used page attribute ``title``.
+        Default behavior is to use page attribute ``title``.
 
         Returns:
             string: Page title.
@@ -192,18 +182,9 @@ class PageViewBase:
         """
         return (len(self.get_destination().split("/")) - 1) * "../" or "./"
 
-    def get_template_name(self):
-        """
-        Get template file path.
-
-        Returns:
-            string: Template file path relative to templates directory.
-        """
-        return self.template_name.format(language_code=self.get_lang().code)
-
     def get_context(self):
         """
-        Get template context.
+        Get view context.
 
         Returns:
             dict: Template context of variables.
@@ -214,7 +195,6 @@ class PageViewBase:
                 "page_destination": self.get_destination(),
                 "page_relative_position": self.get_relative_position(),
                 "page_lang": self.get_lang(),
-                "page_template_name": self.get_template_name(),
             }
         )
 
@@ -224,61 +204,35 @@ class PageViewBase:
 
     def render(self, env):
         """
-        Take the Jinja2 environment as required argument.
+        Base rendering method does not render anything and always return an empty
+        string. You will have to implement the render method yourself or see
+        ``PageTemplateView`` instead if you just want to render a template.
+
+        Although it is does not implement any template logic, this method set the
+        Jinja environment.
 
         Arguments:
             env (jinja2.Jinja2Environment): Jinja environment.
 
         Returns:
-            string: HTML builded from page template with its context.
+            string: An empty string.
         """
         self.env = env
-        context = self.get_context()
 
-        template = self.env.get_template(self.get_template_name())
-
-        return template.render(lang=self.get_lang(), **context)
-
-    def _recurse_template_search(self, env, template_name):
-        """
-        Load involved template sources from given template file path then find
-        their template references.
-
-        Arguments:
-            env (jinja2.Jinja2Environment): Jinja environment.
-            template_name (string): Template file path.
-
-        Returns:
-            list: List of involved templates sources files.
-        """
-        template_source = env.loader.get_source(env, template_name)[0]
-        parsed_content = env.parse(template_source)
-
-        deps = []
-        for item in Jinja2Meta.find_referenced_templates(parsed_content):
-            deps.append(item)
-            deps += self._recurse_template_search(env, item)
-
-        return deps
+        return ""
 
     def introspect(self, env):
         """
-        Take the Jinja2 environment as required argument to find every
-        templates dependancies from page.
+        Dummy introspect method for base view required for internal code like watchers
+        that may use it.
+
+        This does not implement any introspection since there is not template logic
+        here.
 
         Arguments:
             env (jinja2.Jinja2Environment): Jinja environment.
 
         Returns:
-            list: List of involved templates sources files.
+            list: Empty list.
         """
-        if self._used_templates is None:
-            self.env = env
-
-            found = self._recurse_template_search(env, self.get_template_name())
-
-            self._used_templates = [self.get_template_name()] + found
-
-            self.logger.debug(" - Used templates: {}".format(self._used_templates))
-
-        return self._used_templates
+        return []