+++ tests/1000_interfaces/04_watcher.py
@@ -5,10 +5,10 @@ from optimus.interfaces.build import builder_interface
 from optimus.interfaces.starter import starter_interface
 from optimus.interfaces.watch import watcher_interface
 from optimus.logs import set_loggers_level
-from optimus.pages.views.base import PageViewBase
+from optimus.pages.views.base import PageTemplateView
 
 
-class DummyView(PageViewBase):
+class DummyView(PageTemplateView):
     """
     A dummy view similar to the one from "basic" starter.
     """